relocated components
