+++ src/pages/HomePage/components/AuthModal/Sign.scss
@@ -1,4 +1,4 @@
-@import '../../variables';
+@import '../../../../variables';
 
 .content {
   &.authentication {
